anomaly-detector/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
prefect>=3.0.0
pymongo>=4.6.0
python-dotenv>=1.0.0
pydantic>=2.6.0
